+++ src/components/authenticatormodal/index.ts
@@ -1,0 +1,1 @@
+export * from './AuthenticatorModal'
