+++ src/features/daily-nutrition/components/NutritionRing.tsx
@@ -1,5 +1,5 @@
 import { Grid, RingProgress, Text } from '@mantine/core'
-import { Nutrition } from './types'
+import { Nutrition } from '../types'
 
 export function NutritionRing({ name, number, percent, color }: Nutrition) {
   const unit = name === 'Calorie' ? 'Kcal' : 'g'
