+++ src/pages/_app.tsx
@@ -2,7 +2,7 @@ import type { AppProps } from 'next/app'
 import { Authenticator, useAuthenticator } from '@aws-amplify/ui-react'
 import '@/styles/globals.css'
 import '@aws-amplify/ui-react/styles.css'
-import { LandingPage } from '../components/landingpage/LandingPage'
+import { LandingPage } from '../components/LandingPage'
 
 import { Amplify } from 'aws-amplify'
 import awsExports from '../aws-exports'
