+++ src/pages/settings.tsx
@@ -1,4 +1,4 @@
-import { Layout } from '../components/layout/Layout'
+import { Layout } from '../components/Layout'
 
 export default function Settings() {
   return <Layout title="Settings">Settings</Layout>
